refactor deployment, containerization except for consumer and model.
