+++ backend/.env
@@ -7,3 +7,4 @@ SECRET_KEY=you-will-never-guess-so
 SQLALCHEMY_DATABASE_URI=sqlite:///mediainsight.db
 SQLALCHEMY_TRACK_MODIFICATIONS=False
 LOG_TO_STDOUT=True
+KAFKA_BOOTSTRAP_SERVERS='localhost:9092'
