+++ backend/Dockerfile
@@ -1,0 +1,22 @@
+# Use the official Python image as a base image
+FROM python:3.12-slim
+
+# Set the working directory
+WORKDIR /app
+
+# Copy the requirements.txt file and install dependencies
+COPY requirements.txt /app/
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy the rest of the application code
+COPY . /app
+
+# Expose the port that Flask runs on
+EXPOSE 5000
+
+# Set environment variables to ensure Flask runs properly
+ENV FLASK_APP=mediainsight.py
+ENV FLASK_RUN_HOST=0.0.0.0
+
+# Run the application
+CMD ["flask", "run"]
