+++ backend/app/main/__init__.py
@@ -1,2 +1,1 @@
 from app.main import rest_resources
-from app.main import kafka_consumer
