+++ frontend/mediainsight_ui/package.json
@@ -35,5 +35,5 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:5000"
+  "proxy": "http://web:5000"
 }
